spotify playlist can be imported

+++ Kyoku/setup/presentation/src/main/java/com/poulastaa/setup/presentation/spotify_playlist/ImportPlaylistUiState.kt
@@ -22,5 +22,5 @@ data class UiPrevPlaylist(
 data class UiPlaylist(
     val id: Long = -1,
     val name: String = "",
-    val totalSongs: Long = 0,
+    val totalSongs: Int = 0,
 )
